AI_lib/aidk/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
argcomplete==3.6.3
asteval==0.9.33
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
bitsandbytes==0.49.0
boto==2.49.0
cachetools==5.5.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
clarabel==0.11.1
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.2
crcmod==1.7
cryptography==43.0.3
cvxpy==1.7.5
cycler==0.12.1
exceptiongroup==1.3.1
fasteners==0.20
filelock==3.20.1
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
gast==0.7.0
gcs-oauth2-boto-plugin==3.3
google-apitools==0.5.35
google-auth==2.39.0
google-auth-httplib2==0.3.0
google-pasta==0.2.0
google-reauth==0.1.1
grpcio==1.76.0
gsutil==5.35
gym==0.26.2
gym-notices==0.1.0
h5py==3.15.1
hf-xet==1.2.0
httplib2==0.20.4
huggingface-hub==0.36.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
keras==3.12.0
kiwisolver==1.4.9
lava-dl==0.6.0
lava-nc==0.10.0
learn2learn==0.2.0
libclang==18.1.1
linecache2==1.0.0
llvmlite==0.46.0
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mixture-of-experts==0.2.3
ml_dtypes==0.5.4
monotonic==1.6
mpmath==1.3.0
multidict==6.7.0
namex==0.1.0
networkx==2.8.7
ninja==1.13.0
nltk==3.9.2
numba==0.63.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.1.105
oauth2client==4.1.3
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.18.0
osqp==1.0.5
packaging==25.0
pandas==2.3.3
pillow==10.4.0
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
pyOpenSSL==24.2.1
pyparsing==3.3.1
pytest==7.4.4
python-dateutil==2.9.0.post0
pytz==2025.2
pyu2f==0.1.5
PyYAML==6.0.3
qpth==0.0.18
regex==2025.11.3
requests==2.32.5
retry-decorator==1.1.1
rich==14.2.0
rsa==4.7.2
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
scs==3.2.10
sentence-transformers==5.2.0
sentencepiece==0.2.1
six==1.17.0
sympy==1.13.1
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.3.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
traceback2==1.4.0
transformers==4.57.3
triton==3.1.0
typing_extensions==4.15.0
tzdata==2025.3
unittest2==1.1.0
urllib3==2.6.2
Werkzeug==3.1.4
wrapt==2.0.1
yarl==1.22.0
